requirements.txt
pyyaml
fade
colorama
pystyle